Refacout out engines eventlisteners to separate class
